fix bug

+++ library/src/main/java/com/dd/morphingbutton/impl/LinearProgressButton.java
@@ -6,6 +6,7 @@ import android.graphics.Paint;
 import android.graphics.RectF;
 import android.support.annotation.NonNull;
 import android.util.AttributeSet;
+
 import com.dd.morphingbutton.IProgress;
 import com.dd.morphingbutton.MorphingButton;
 
@@ -48,7 +49,7 @@ public class LinearProgressButton extends MorphingButton implements IProgress {
                 mRectF = new RectF();
             }
 
-            mRectF.right = (getWidth() / MAX_PROGRESS) * mProgress;
+            mRectF.right = ((float) getWidth() / MAX_PROGRESS) * mProgress;
             mRectF.bottom = getHeight();
             canvas.drawRoundRect(mRectF, mProgressCornerRadius, mProgressCornerRadius, mPaint);
         }
